refactor: handler, parser, validator 패키지 분리
- handler 패키지: InputHandler, OutputHandler
- parser 패키지: StringParser, DelimiterExtractor,
DelimiterExtractionResult
- validator 패키지: InputValidator
