+++ src/main/java/calculator/handler/OutputHandler.java
@@ -1,11 +1,11 @@
-package calculator;
+package calculator.handler;
 
 public class OutputHandler {
-    void printInitMessage() {
+    public void printInitMessage() {
         System.out.println("덧셈할 문자열을 입력해 주세요.");
     }
 
-    void printResult(int result) {
+    public void printResult(int result) {
         System.out.println("결과 : " + result);
     }
 }
