+++ src/main/java/calculator/parser/DelimiterExtractor.java
@@ -1,5 +1,6 @@
-package calculator;
+package calculator.parser;
 
+import calculator.parser.dto.DelimiterExtractionResult;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.regex.Matcher;
