+++ src/main/java/calculator/parser/StringParser.java
@@ -1,5 +1,7 @@
-package calculator;
+package calculator.parser;
 
+import calculator.parser.dto.DelimiterExtractionResult;
+import calculator.validator.InputValidator;
 import java.util.Arrays;
 import java.util.List;
 import java.util.regex.Pattern;
