+++ src/main/java/calculator/parser/dto/DelimiterExtractionResult.java
@@ -1,11 +1,11 @@
-package calculator;
+package calculator.parser.dto;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class DelimiterExtractionResult {
-    String input;
-    List<String> delimiters = new ArrayList<>(List.of(",", ":"));
+    public String input;
+    public List<String> delimiters = new ArrayList<>(List.of(",", ":"));
 
     public DelimiterExtractionResult(String input, List<String> delimiters) {
         this.input = input;
