+++ src/main/java/calculator/validator/InputValidator.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.validator;
 
 public class InputValidator {
     public static String validateToken(String token) {
